mononoke: split ChangesetPathContext
Summary:
Previously ChangesetPathContext was holding both fsnode_id and unode_id, however it made it easier to misuse api and trigger expensive fsnodes or unodes path traversal (see more info in the comments for D27587664).

This diff splits it in two separate types.
Also I noticed that ChangesetPathContext wasn't using the `unode_id` future that it stored, so I just deleted it.

Reviewed By: markbt

Differential Revision: D27590997

fbshipit-source-id: 08fc14d33c82357275413c4cf2698f97620503ea