mononoke/lfs_server: don't trust the client to provide valid size
Summary:
Historically, we haven't really cared about the sizes provided by clients in
LFS, and we went as far as just echoing them back to the client.

However, with the support I'm adding for range requests, this will start to
matter, because clients will ask for content in this range, so if the client
has the wrong size, we should correct it for them rather than just let them
proceed, lest they fail to download the file properly.

That being said, I am pretty sure there are places relying on us not caring
about the size, so I'm not throwing errors there.

Reviewed By: mitrandir77

Differential Revision: D27710438

fbshipit-source-id: ab670b44364604c07c449e500e379ca40b8c5ec1