mononoke/filestore: don't take a chunk when you need 0 bytes from it
Summary:
If you ask for a range starting at byte 3, and the chunks are of size 3, then
you don't actually need need the first chunk, but right now we'll fetch it
then extract zero bytes from it.

This is quite wasteful, and for LFS range fetches will be problematic since it
basically doubles the volume of stuff we need to keep in cache (we need both
chunks).

Reviewed By: farnz

Differential Revision: D27824411

fbshipit-source-id: 7103f5b4d5bb78f023245f3e8a1bcb0c2f28faab