mononoke: add filenodes to hg warmers
Summary:
Let's make sure filenodes are warmed at the same time as hg changesets.
We do it implicitly by deriving them in getbundle, however it results in more
attempts of deriving filenodes, because we'd try to derive on every pull.

This is not a huge problem, but that would mean that we could send more
requests to the root filenode to check whether filenodes for a commit were derived or not.
And I remember looking at the logs a few days ago and seeing a few filenodes errors in our
scuba table, which I suspected might be related to the fact that filenodes are
not in derived data cache.

There's a high chance i'm wrong about the root cause of these problems this change should be
reasonable to do anyway.

Reviewed By: krallin

Differential Revision: D27822519

fbshipit-source-id: c878206a09bd19e4a68327dddf27aae10490740f